impl(bigquery): move Interval and Range to datatypes mod
